Only use first intersection with the terrain
I have observed that there might be more than just one intersection when doing raycasting to get the distance of the map from the camera. I guess that only the first intersection should be used to determine the LOD, right?

This situation of multiple intersections during one raycast happend when the camera was close to the ground and looking towards mountains (MapView with MapView.HEIGHT).